Update README.md
An interesting mind game help to develop memory for kids as well as a youngster using HTML, CSS, JavaScript in the front-end. It is a game that starts with a keypress and continues till you remember the pattern in which the events (blinking of color dice ) occurred and press that color dice in the same pattern as it occurred. It comprises mainly of Javascript functions and event handling functions along with few codes of HTML and CSS.So go for it and show your memorizing power how far you can go !! BEST OF LUCK!!